fastapi/requirements.txt
fastapi
uvicorn
scikit-learn==0.20.2
profanity-check==1.0.2
redis
cassandra-driver
pika
requests 
